+++ Nos/Views/Components/CopyButtonState.swift
@@ -1,0 +1,5 @@
+/// The state of a copy button.
+enum CopyButtonState {
+    case copy
+    case copied
+}